refactor(storage): extract ProfileSearchMixin from ProfileMixin (SQLite + ABC)
Final extraction of the _profiles.py decomposition. Move the Search bucket
(search_interaction, search_user_profile) verbatim into a new
profiles/_search.py::ProfileSearchMixin on both the SQLite backend and the ABC
base, and compose it into SQLiteStorage/BaseStorage alongside ProfileStoreMixin
and InteractionStoreMixin.

The shared _build_tags_sql helper stays module-level in the residual _profiles.py
(used by both ProfileStore and ProfileSearch) and is imported into _search.py.
The residual ProfileMixin is now drained of all public methods (Task 6 removes
the class and its composition).
