+++ reflexio/server/services/storage/sqlite_storage/profiles/__init__.py
@@ -1,7 +1,9 @@
 from ._interaction_store import InteractionStoreMixin
 from ._profile_store import ProfileStoreMixin
+from ._search import ProfileSearchMixin
 
 __all__ = [
     "InteractionStoreMixin",
+    "ProfileSearchMixin",
     "ProfileStoreMixin",
 ]
